requirements.txt
pyserial
opencv-python-headless
numpy
pandas
tensorflow
keras-tcn
scikit-learn